requirements.txt
customtkinter
python-Levenshtein
